docs: update admonition.mdx
